+++ src/constants.ts
@@ -1,0 +1,15 @@
+export enum CommandCategory {
+    SUPPORT = 'Support',
+    GENERAL = 'General',
+    UTILS = 'Utilities',
+    MEMES = 'Memes',
+    MODERATION = 'Moderation',
+}
+
+export enum ResponseType {
+    STATIC,
+    EPHEMERAL,
+    EDIT,
+    DEFER,
+    FOLLOW_UP,
+}