+++ src/index.ts
@@ -1,7 +1,10 @@
-import { Partials, Client } from 'discord.js';
-import * as dotenv from 'dotenv'
+import {Client, Partials } from 'discord.js';
+import dotenv from 'dotenv';
+import Logger, { createCommand} from './lib';
+import {commands} from './commands/index';
 
 dotenv.config();
+
 export const client = new Client({
     intents: [
         'Guilds',
@@ -10,7 +13,6 @@ export const client = new Client({
         'GuildMessages',
         'GuildMessageReactions',
         'GuildEmojisAndStickers',
-        'GuildBans',
         'DirectMessages',
         'DirectMessageReactions',
         'DirectMessageTyping',
@@ -19,21 +21,52 @@ export const client = new Client({
     partials: [
         Partials.Channel,
         Partials.Message,
-        Partials.User,
         Partials.GuildMember,
         Partials.Reaction,
+        Partials.User,
     ],
 });
 
-client.on("ready", () => {
-    console.log("Bot is online");
+client.login(process.env.DISCORD_TOKEN)
+    .then()
+    .catch((e) => {
+        Logger.error(e);
+        process.exit(1);
+    });
+
+client.on('ready', () => {
+    if(client.user == null) {
+        console.log('error, client not found');
+    } else {
+    console.log(`Logged in as ${client.user.username}`)
+    }
+
+    for (const command of commands) {
+        const slashCommand = createCommand(command);
+        client.application?.commands.create(slashCommand);
+    }
+
 });
 
-client.on("messageCreate", (message) => {
-    if (message.content.startsWith("ping")) {
-        message.channel.send("pong!");
+client.on('interactionCreate', async (interaction: any) => {
+    if (!interaction.isCommand()) return;
+
+    if(!interaction.guild) {
+        Logger.error('bailing because interaction in DM');
+        return;
+    }
+
+    if(interaction.user.bot) {
+        Logger.error('Bailing due to bot message');
+    }
+
+    for(const command of commands) {
+        if (interaction.commandName == command.name) {
+            if(command.interaction == undefined) {
+                Logger.error(`Error: ${command.name} has no interaction`);
+                break;
+            }
+                await command.interaction(interaction)
+        }
     }
 });
-
-
-client.login(process.env.DISCORD_TOKEN);