+++ src/lib/CommandOptions.ts
@@ -1,0 +1,4 @@
+export interface InputCommandOptions {
+    name: string,
+    description: string
+}