+++ src/lib/index.ts
@@ -1,0 +1,5 @@
+export * from './logger';
+export * from './CommandDefinition';
+export * from './CommandOptions';
+export * from './createCommand';
+export { default } from './logger';