+++ .github/CHANGELOG.md
@@ -1,2 +1,3 @@
 # November 2022
 - feat: (11/04/2022): Initialize new discord.js bot
+- feat: (03/07/2023): Integrate slash command api & new cowsay command